added validation for step one
