+++ src/api/TokenStatsApi.ts
@@ -1,10 +1,11 @@
 import { useNetwork } from "@contexts/NetworkContext";
+import { NetworkConnection } from "@contexts/Environment";
 import mockData from "../mockdata/TokenPriceHistory";
 
 export default {
   useTokenStats: () => {
     const { connection } = useNetwork();
-    if (connection === "MainNet") {
+    if (connection === NetworkConnection.MainNet) {
       // TODO: This condition is for devs to test negative trend data. Remove once api is ready.
       return {
         ...mockData.tokenStats,
@@ -15,7 +16,7 @@ export default {
   },
   useTokenPriceHistory: () => {
     const { connection } = useNetwork();
-    if (connection === "MainNet") {
+    if (connection === NetworkConnection.MainNet) {
       // TODO: This condition is for devs to test negative trend data. Remove once api is ready.
       return [...mockData.tokenPriceChangeHistory].reverse();
     }
