+++ src/components/commons/InputComponent.tsx
@@ -16,18 +16,24 @@ export default function InputComponent({
   placeholder: string;
 }) {
   const [isFocused, setIsFocused] = useState<boolean>(false);
+  const [isVisited, setIsVisited] = useState<boolean>(false);
   return (
     <div>
       <div className="text-lg mb-2 text-white-700">{label}</div>
       <div
         className={clsx(
           "flex flex-row justify-between px-4 py-[18px] bg-dark-100 rounded-[10px] border-[0.5px] border-dark-200 focus-within:border focus-within:border-white-900",
-          { "!border border-red-800": error }
+          { "!border !border-red-800": error && isVisited }
         )}
       >
         <input
-          onFocus={() => setIsFocused(true)}
-          onBlur={() => setIsFocused(false)}
+          onFocus={() => {
+            setIsFocused(true);
+          }}
+          onBlur={() => {
+            setIsFocused(false);
+            setIsVisited(true);
+          }}
           className="w-full focus:outline-none border-none bg-dark-100 text-white-50 text-xl placeholder-black-500"
           onChange={(v) => setValue(v.target.value)}
           placeholder={placeholder}
